+++ CHANGES.md
@@ -1,6 +1,7 @@
 0.1.1 (unreleased)
 =====
 
+- Allow specifying both a filename and a regexp.
 - Correctly count questions answered questions with 0 points.
 
 0.1.0 (2025-10-10)